Create generic PrimaryAlert dialog
